Fix a segfault by initializing attr_dict to NULL
I thought PyObject_New would call my tp_new, which is set to
PyType_GenericNew, which NULLs out all user-defined fields (actually
memset, but the docs say NULL).  However, this appears to not be the
case.  Explicitly NULL out attr_dict in create_regexp to avoid segfaults
when compiling a bad pattern.  Do it for create_match too for future
safety, even though it is not required with the current code.