Performance improvements for pyre2
As part of a quick test I did to measure the perfomance of pyre2, I noticed
thathat we spent a lot of time allocating objects and that we could improve
thethe performance simply by altering some of the C code to limit unnecessary
object allocations. The main culprit was Py_BuildValue but it seems like the
changes had a positive effect overall in terms of performance.

As part of that I did:

* Deleted code that populated attributes of the __dict__ struct
* Deleted code that initialized the struct itself